Refactor tests: Consolidate builder tests for BBox2D and BBox3D into main test files
- Removed separate builder test files for BBox2D and BBox3D.
- Integrated builder tests into existing BBox2D and BBox3D test files.
- Merged transformation tests for Point2D and Point3D into their respective test files.
- Cleaned up unused test modules and ensured all tests are properly organized.
