+++ geo_primitives/src/point_2d_tests.rs
@@ -1,4 +1,4 @@
-﻿//! Point2D の包括的テスト
+//! Point2D の包括的テスト
 //!
 //! 基本機能、座標操作、距離計算、変換機能、演算子などをテスト
 
@@ -299,5 +299,61 @@ mod tests {
         // 距離の計算
         let norm = neg_point.distance_to(&Point2D::<f64>::origin());
         assert!((norm - 5.0).abs() < 1e-10);
+    }
+
+    // ============================================================================
+    // Transform テスト (point_2d_transform.rs から統合)
+    // ============================================================================
+
+    #[test]
+    fn test_transform_translate() {
+        use geo_foundation::extensions::BasicTransform;
+
+        let point = Point2D::new(1.0, 2.0);
+        let translation = Vector2D::new(2.0, 3.0);
+        let translated = BasicTransform::translate(&point, translation);
+
+        assert_eq!(translated.x(), 3.0);
+        assert_eq!(translated.y(), 5.0);
+    }
+
+    #[test]
+    fn test_transform_scale_origin() {
+        use geo_foundation::extensions::TransformHelpers;
+
+        let point = Point2D::new(2.0, 3.0);
+        let scaled = point.scale_origin(2.0);
+
+        assert_eq!(scaled.x(), 4.0);
+        assert_eq!(scaled.y(), 6.0);
+    }
+
+    #[test]
+    fn test_transform_translate_axes() {
+        use geo_foundation::extensions::TransformHelpers;
+
+        let point = Point2D::new(1.0, 2.0);
+
+        let translated_x = point.translate_x(1.0);
+        assert_eq!(translated_x, Point2D::new(2.0, 2.0));
+
+        let translated_y = point.translate_y(1.0);
+        assert_eq!(translated_y, Point2D::new(1.0, 3.0));
+
+        let translated_xy = point.translate_xy(1.0, 1.0);
+        assert_eq!(translated_xy, Point2D::new(2.0, 3.0));
+    }
+
+    #[test]
+    fn test_transform_scale_from_center() {
+        use geo_foundation::extensions::BasicTransform;
+
+        let point = Point2D::new(4.0, 6.0);
+        let center = Point2D::new(2.0, 3.0);
+        let scaled = BasicTransform::scale(&point, center, 2.0);
+
+        // (4-2)*2+2=6, (6-3)*2+3=9
+        assert_eq!(scaled.x(), 6.0);
+        assert_eq!(scaled.y(), 9.0);
     }
 }
