+++ geo_primitives/src/point_2d_transform.rs
@@ -64,57 +64,3 @@ impl<T: Scalar> BasicTransform<T> for Point2D<T> {
 // Required implementations for BasicTransform
 // Note: Default trait is already implemented in point_2d.rs
 // ============================================================================
-
-// ============================================================================
-// Tests Module
-// ============================================================================
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use geo_foundation::extensions::TransformHelpers;
-
-    #[test]
-    fn test_translate() {
-        let point = Point2D::new(1.0, 2.0);
-        let translation = Vector2D::new(2.0, 3.0);
-        let translated = point.translate(translation);
-
-        assert_eq!(translated.x(), 3.0);
-        assert_eq!(translated.y(), 5.0);
-    }
-
-    #[test]
-    fn test_scale_origin() {
-        let point = Point2D::new(2.0, 3.0);
-        let scaled = point.scale_origin(2.0);
-
-        assert_eq!(scaled.x(), 4.0);
-        assert_eq!(scaled.y(), 6.0);
-    }
-
-    #[test]
-    fn test_translate_axes() {
-        let point = Point2D::new(1.0, 2.0);
-
-        let translated_x = point.translate_x(1.0);
-        assert_eq!(translated_x, Point2D::new(2.0, 2.0));
-
-        let translated_y = point.translate_y(1.0);
-        assert_eq!(translated_y, Point2D::new(1.0, 3.0));
-
-        let translated_xy = point.translate_xy(1.0, 1.0);
-        assert_eq!(translated_xy, Point2D::new(2.0, 3.0));
-    }
-
-    #[test]
-    fn test_scale_from_center() {
-        let point = Point2D::new(4.0, 6.0);
-        let center = Point2D::new(2.0, 3.0);
-        let scaled = BasicTransform::scale(&point, center, 2.0);
-
-        // (4-2)*2+2=6, (6-3)*2+3=9
-        assert_eq!(scaled.x(), 6.0);
-        assert_eq!(scaled.y(), 9.0);
-    }
-}
