+++ geo_primitives/src/point_3d_transform.rs
@@ -109,60 +109,3 @@ impl<T: Scalar> FromAnalysisVector3<T> for Point3D<T> {
         Point3D::new(vec.x(), vec.y(), vec.z())
     }
 }
-
-// ============================================================================
-// Tests Module
-// ============================================================================
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use geo_foundation::extensions::TransformHelpers;
-
-    #[test]
-    fn test_translate() {
-        let point = Point3D::new(1.0, 2.0, 3.0);
-        let translation = Vector3D::new(2.0, 3.0, 4.0);
-        let translated = point.translate(translation);
-
-        assert_eq!(translated.x(), 3.0);
-        assert_eq!(translated.y(), 5.0);
-        assert_eq!(translated.z(), 7.0);
-    }
-
-    #[test]
-    fn test_scale_origin() {
-        let point = Point3D::new(2.0, 3.0, 4.0);
-        let scaled = point.scale_origin(2.0);
-
-        assert_eq!(scaled.x(), 4.0);
-        assert_eq!(scaled.y(), 6.0);
-        assert_eq!(scaled.z(), 8.0);
-    }
-
-    #[test]
-    fn test_translate_axes() {
-        let point = Point3D::new(1.0, 2.0, 3.0);
-
-        let translated_x = point.translate_x(1.0);
-        assert_eq!(translated_x, Point3D::new(2.0, 2.0, 3.0));
-
-        let translated_y = point.translate_y(1.0);
-        assert_eq!(translated_y, Point3D::new(1.0, 3.0, 3.0));
-
-        let translated_xy = point.translate_xy(1.0, 1.0);
-        assert_eq!(translated_xy, Point3D::new(2.0, 3.0, 3.0));
-    }
-
-    #[test]
-    fn test_scale_from_center() {
-        let point = Point3D::new(4.0, 6.0, 8.0);
-        let center = Point3D::new(2.0, 3.0, 4.0);
-        let scaled = point.scale(&center, 2.0);
-
-        // (4-2)*2+2=6, (6-3)*2+3=9, (8-4)*2+4=12
-        assert_eq!(scaled.x(), 6.0);
-        assert_eq!(scaled.y(), 9.0);
-        assert_eq!(scaled.z(), 12.0);
-    }
-}
